server/requirements.txt
attrs==20.3.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-MySQL==1.5.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
linecache2==1.0.0
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
traceback2==1.4.0
unittest2==1.1.0
Werkzeug==1.0.1
